refactor: moved into address_resolver
